refactor: Implement deeper structural improvements
This commit implements a second round of structural improvements across various areas of the codebase:

- **Code Organization (packages/cli/src/commands/extensions/install.ts):**
  Moved the `getInstallMetadata` function and `InstallArgs` interface to a new dedicated module `packages/cli/src/config/extension-source-resolver.ts` for enhanced modularity.

- **Overall Repository Layout (scripts/):**
  Extracted the `copyFilesRecursive` utility into a new shared module `scripts/utils/file-copy.js` and updated `scripts/copy_files.js` to use it, improving script modularity and reusability.

- **Dependency Management (package.json):**
  Continued standardizing dependency versions in the root `package.json` to exact versions for improved reproducibility, covering numerous `@types/*`, `vitest/*`, `eslint*`, `glob`, `globals`, `google-artifactregistry-auth`, `husky`, `json`, `lint-staged`, `memfs`, `mnemonist`, `mock-fs`, `msw`, `npm-run-all`, `prettier`, `react-devtools-core`, `semver`, `tsx`, `typescript-eslint`, `vitest`, `yargs`, `simple-git`, and `node-pty` entries.

- **Build Process (.github/workflows/ci.yml):**
  Added an `if` condition to the `test_linux` job in `ci.yml` to ensure it runs only when the workflow is triggered by relevant events, contributing to more granular CI control.

- **React Guidelines (packages/cli/src/ui/components/Composer.tsx):**
  Extracted a new `ComposerStatusDisplay.tsx` component from `Composer.tsx` to further decompose the UI, reducing complexity and improving readability of the main `Composer` component.
