+++ .github/workflows/ci.yml
@@ -86,6 +86,10 @@ jobs:
   test_linux:
     name: 'Test (Linux)'
     runs-on: 'gemini-cli-ubuntu-16-core'
+    if: |
+      github.event_name == 'workflow_dispatch' ||
+      github.event_name == 'pull_request' ||
+      github.event_name == 'push'
     needs:
       - 'lint'
     permissions:
