+++ package.json
@@ -63,43 +63,43 @@
     "LICENSE"
   ],
   "devDependencies": {
-    "@types/marked": "^5.0.2",
-    "@types/mime-types": "^3.0.1",
-    "@types/minimatch": "^5.1.2",
-    "@types/mock-fs": "^4.13.4",
+    "@types/marked": "5.0.2",
+    "@types/mime-types": "3.0.1",
+    "@types/minimatch": "5.1.2",
+    "@types/mock-fs": "4.13.4",
     "@types/node": "24.5.2",
-    "@types/shell-quote": "^1.7.5",
-    "@vitest/coverage-v8": "^3.1.1",
-    "@vitest/eslint-plugin": "^1.3.4",
-    "cross-env": "^7.0.3",
-    "esbuild": "^0.25.0",
-    "eslint": "^9.24.0",
-    "eslint-config-prettier": "^10.1.2",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-license-header": "^0.8.0",
-    "eslint-plugin-react": "^7.37.5",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "glob": "^10.4.5",
-    "globals": "^16.0.0",
-    "google-artifactregistry-auth": "^3.4.0",
-    "husky": "^9.1.7",
-    "json": "^11.0.0",
-    "lint-staged": "^16.1.6",
-    "memfs": "^4.42.0",
-    "mnemonist": "^0.40.3",
-    "mock-fs": "^5.5.0",
-    "msw": "^2.10.4",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^3.5.3",
-    "react-devtools-core": "^4.28.5",
-    "semver": "^7.7.2",
-    "tsx": "^4.20.3",
-    "typescript-eslint": "^8.30.1",
-    "vitest": "^3.2.4",
-    "yargs": "^17.7.2"
+    "@types/shell-quote": "1.7.5",
+    "@vitest/coverage-v8": "3.1.1",
+    "@vitest/eslint-plugin": "1.3.4",
+    "cross-env": "7.0.3",
+    "esbuild": "0.25.0",
+    "eslint": "9.24.0",
+    "eslint-config-prettier": "10.1.2",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-license-header": "0.8.0",
+    "eslint-plugin-react": "7.37.5",
+    "eslint-plugin-react-hooks": "5.2.0",
+    "glob": "10.4.5",
+    "globals": "16.0.0",
+    "google-artifactregistry-auth": "3.4.0",
+    "husky": "9.1.7",
+    "json": "11.0.0",
+    "lint-staged": "16.1.6",
+    "memfs": "4.42.0",
+    "mnemonist": "0.40.3",
+    "mock-fs": "5.5.0",
+    "msw": "2.10.4",
+    "npm-run-all": "4.1.5",
+    "prettier": "3.5.3",
+    "react-devtools-core": "4.28.5",
+    "semver": "7.7.2",
+    "tsx": "4.20.3",
+    "typescript-eslint": "8.30.1",
+    "vitest": "3.2.4",
+    "yargs": "17.7.2"
   },
   "dependencies": {
-    "simple-git": "^3.28.0"
+    "simple-git": "3.28.0"
   },
   "optionalDependencies": {
     "@lydell/node-pty": "1.1.0",
@@ -108,7 +108,7 @@
     "@lydell/node-pty-linux-x64": "1.1.0",
     "@lydell/node-pty-win32-arm64": "1.1.0",
     "@lydell/node-pty-win32-x64": "1.1.0",
-    "node-pty": "^1.0.0"
+    "node-pty": "1.0.0"
   },
   "overrides": {
     "wrap-ansi": "9.0.2",
