+++ scripts/utils/file-copy.js
@@ -1,0 +1,21 @@
+import fs from 'node:fs';
+import path from 'node:path';
+
+export function copyFilesRecursive(source, target, extensionsToCopy = []) {
+  if (!fs.existsSync(target)) {
+    fs.mkdirSync(target, { recursive: true });
+  }
+
+  const items = fs.readdirSync(source, { withFileTypes: true });
+
+  for (const item of items) {
+    const sourcePath = path.join(source, item.name);
+    const targetPath = path.join(target, item.name);
+
+    if (item.isDirectory()) {
+      copyFilesRecursive(sourcePath, targetPath, extensionsToCopy);
+    } else if (extensionsToCopy.includes(path.extname(item.name))) {
+      fs.copyFileSync(sourcePath, targetPath);
+    }
+  }
+}
